Made ModGore derive from ModTexturedType, added GoreLoader.
