+++ patches/tModLoader/Terraria/ModLoader/Mod.Internals.cs
@@ -93,7 +93,6 @@ namespace Terraria.ModLoader
 			IList<Type> modGores = new List<Type>();
 			IList<Type> modSounds = new List<Type>();
 
-
 			Type modType = GetType();
 			foreach (Type type in Code.GetTypes().OrderBy(type => type.FullName, StringComparer.InvariantCulture)) {
 				if (type == modType){continue;}
@@ -114,7 +113,7 @@ namespace Terraria.ModLoader
 				}
 			}
 			if (Properties.AutoloadGores) {
-				AutoloadGores(modGores);
+				GoreLoader.AutoloadGores(this, modGores);
 			}
 			if (Properties.AutoloadSounds) {
 				AutoloadSounds(modSounds);
@@ -178,18 +177,6 @@ namespace Terraria.ModLoader
 		private void AutoloadBackgrounds() {
 			foreach (string texture in Assets.EnumeratePaths<Texture2D>().Where(t => t.StartsWith("Backgrounds/"))) {
 				AddBackgroundTexture($"{Name}/{texture}");
-			}
-		}
-
-		private void AutoloadGores(IList<Type> modGores) {
-			var modGoreNames = modGores.ToDictionary(t => t.FullName);
-
-			foreach (string texturePath in Assets.EnumeratePaths<Texture2D>().Where(t => t.StartsWith("Gores/"))) {
-				ModGore modGore = null;
-				if (modGoreNames.TryGetValue($"{Name}.{texturePath.Replace('/', '.')}", out Type t))
-					modGore = (ModGore)Activator.CreateInstance(t);
-
-				AddGore($"{Name}/{texturePath}", modGore);
 			}
 		}
